refactor: Split SimpleDashboard.tsx into focused components
Reduce web/src/pages/SimpleDashboard.tsx from 640 lines to 184 lines
by extracting data, helpers, and UI sections into reusable modules.

Extracted modules:
- data/dashboardPresets.ts: TEAM_PRESETS, TEST_SCENARIOS, and types
- utils/dashboardRandom.ts: generateRandomScenario, generateRandomAgents
- components/dashboard/HeroSection.tsx: hero banner
- components/dashboard/TeamSelector.tsx: team preset picker
- components/dashboard/ScenarioSelector.tsx: scenario preset picker
- components/dashboard/AboutSection.tsx: collapsible info panel

The page now orchestrates state with useCallback / useMemo and composes
the smaller components. All subcomponents are wrapped in React.memo.
No behavior changes.
