+++ web/src/components/dashboard/TeamSelector.tsx
@@ -1,0 +1,65 @@
+import React from 'react';
+import { Shuffle } from 'lucide-react';
+import type { TeamPreset } from '../../data/dashboardPresets';
+
+interface TeamSelectorProps {
+  teams: TeamPreset[];
+  selectedId: string;
+  onSelect: (id: string) => void;
+  onRandom: () => void;
+}
+
+/**
+ * Card listing team presets with a radio-style picker.
+ *
+ * Receives the team list, the selected id, and callbacks for selection and
+ * "pick a random preset" so it stays free of dashboard-specific state.
+ */
+export const TeamSelector: React.FC<TeamSelectorProps> = React.memo(
+  ({ teams, selectedId, onSelect, onRandom }) => {
+    return (
+      <div className="card flex flex-col h-[500px]">
+        <div className="flex items-center justify-between mb-4 flex-shrink-0">
+          <h2 className="text-xl font-semibold text-gray-900 dark:text-gray-100">Choose Team</h2>
+          <button
+            onClick={onRandom}
+            className="btn-secondary text-sm flex items-center gap-2"
+          >
+            <Shuffle className="w-4 h-4" />
+            Random
+          </button>
+        </div>
+
+        <div className="space-y-3 flex-1 overflow-y-auto min-h-0 pr-2">
+          {teams.map((team) => (
+            <label
+              key={team.id}
+              className={`block p-4 rounded-lg border-2 cursor-pointer transition-all ${
+                selectedId === team.id
+                  ? 'border-blue-500 bg-blue-50 dark:bg-blue-950'
+                  : 'border-gray-200 dark:border-gray-700 hover:border-blue-300 dark:hover:border-blue-600'
+              }`}
+            >
+              <input
+                type="radio"
+                name="team"
+                value={team.id}
+                checked={selectedId === team.id}
+                onChange={(e) => onSelect(e.target.value)}
+                className="sr-only"
+              />
+              <div className="flex items-start">
+                <div className="flex-1">
+                  <div className="font-medium text-gray-900 dark:text-gray-100">{team.name}</div>
+                  <div className="text-sm text-gray-600 dark:text-gray-400 mt-1">{team.description}</div>
+                </div>
+              </div>
+            </label>
+          ))}
+        </div>
+      </div>
+    );
+  }
+);
+
+TeamSelector.displayName = 'TeamSelector';
